+++ docs/_chunks/index-iife-xwj97sz9.js
@@ -1,5 +1,0 @@
-var d=Object.defineProperty;var e=(a)=>a;function f(a,b){this[a]=e.bind(null,b)}var g=(a,b)=>{for(var c in b)d(a,c,{get:b[c],enumerable:!0,configurable:!0,set:f.bind(b,c)})};var h=((a)=>typeof require<"u"?require:typeof Proxy<"u"?new Proxy(a,{get:(b,c)=>(typeof require<"u"?require:b)[c]}):a)(function(a){if(typeof require<"u")return require.apply(this,arguments);throw Error('Dynamic require of "'+a+'" is not supported')});
-export{g as Sa,h as Ta};
-
-//# debugId=37C62D95CDC09B2564756E2164756E21
-//# sourceMappingURL=index-iife-xwj97sz9.js.map